requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
paramiko==3.4.1
python-dotenv==1.0.1
